+++ core/src/models/uniframe_studio/uniframe_studio.rs
@@ -6,7 +6,7 @@ pub struct DubbingPipelinePrepareRequest {
     pub system_file_name: String,
     pub original_file_name: String,
     pub content_type: String,
-    pub video_duration_seconds: u32,
+    pub video_duration_seconds: f64,
 }
 
 #[derive(Debug, Serialize)]
@@ -16,7 +16,7 @@ pub struct DubbingPipelinePrepareResponse {
     pub video_s3_url: String,
     pub expires_in: u64,
     pub estimated_cost_usd: f64,
-    pub video_duration_seconds: u32,
+    pub video_duration_seconds: f64,
 }
 
 #[derive(Debug, Deserialize)]
